pg: Make MPIDI_Process.my_pg_size an integer instead of unsigned int.
Since MPIDI_Process.my_pg_size is always set to the value of
MPIR_Process.size which is an integer, the datatype of my_pg_size
is changed from unsigned int to int for consistency.

+++ mpich2/src/mpid/psp/src/mpid_psp_connect.c
@@ -22,9 +22,9 @@ struct InitMsg {
 static
 void grank2con_set(int dest_grank, pscom_connection_t * con)
 {
-    unsigned int pg_size = MPIDI_Process.my_pg_size;
+    int pg_size = MPIDI_Process.my_pg_size;
 
-    assert((unsigned int) dest_grank < pg_size);
+    assert(dest_grank < pg_size);
 
     MPIDI_Process.grank2con[dest_grank] = con;
 }
@@ -33,9 +33,9 @@ void grank2con_set(int dest_grank, pscom_connection_t * con)
 static
 pscom_connection_t *grank2con_get(int dest_grank)
 {
-    unsigned int pg_size = MPIDI_Process.my_pg_size;
+    int pg_size = MPIDI_Process.my_pg_size;
 
-    assert((unsigned int) dest_grank < pg_size);
+    assert(dest_grank < pg_size);
 
     return MPIDI_Process.grank2con[dest_grank];
 }
